fix: Frontend
Try to fix an error in staging environment
