+++ apps/solutionazer-frontend/solutionazer/src/app/(auth)/profiles/page.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 /**
  * This file is part of solutionazer.app.
  *
@@ -19,93 +18,15 @@
 
 'use client'
 
+import ProfilesClient from '@/components/auth/profiles/pages/ProfilesClient'
+import { Suspense } from 'react'
+
 export const dynamic = 'force-dynamic'
 
-import Card from '@/components/auth/profiles/Card'
-import Article from '@/components/shared/containers/Article'
-import Section from '@/components/shared/containers/Section'
-import MediumTitle from '@/components/shared/titles/MediumTitle'
-import SmallTitle from '@/components/shared/titles/SmallTitle'
-import Title from '@/components/shared/titles/Title'
-import Admin from '@/lib/auth/companies/admins/admin'
-import Company from '@/lib/auth/companies/company'
-import Member from '@/lib/auth/companies/members/member'
-import useAuthStore from '@/lib/auth/states/global/authStore'
-import { getCompaniesByUser } from '@/lib/utils/users-management/companyHandler'
-import { useEffect, useState } from 'react'
-
-import styles from './page.module.css'
-
 export default function Profiles() {
-  // auth global state
-  const { user } = useAuthStore()
-
-  // companies
-  const [companies, setCompanies] = useState<Company[]>([])
-
-  // load companies
-  useEffect(() => {
-    const fetchData = async () => {
-      if (user?.getUuid()) {
-        const data = (await getCompaniesByUser(user.getUuid())).map(
-          (company: {
-            uuid: string
-            name: string
-            admins: Admin[]
-            members: Member[]
-          }) => {
-            return new Company({
-              uuid: company.uuid,
-              name: company.name,
-              admins: company.admins.map((admin: any) => {
-                return new Admin({
-                  uuid: admin.uuid,
-                  fullName: admin.fullName,
-                  email: admin.email,
-                })
-              }),
-              members: company.members.map((member: any) => {
-                return new Member({
-                  uuid: member.uuid,
-                  fullName: member.fullName,
-                  email: member.email,
-                })
-              }),
-            })
-          },
-        )
-
-        console.log(data)
-
-        setCompanies(data)
-      }
-    }
-
-    fetchData()
-  }, [user])
-
   return (
-    <div className={styles.profiles_content}>
-      <Title params={{ text: 'Profiles', classNames: ['title'] }} />
-      <Section>
-        <MediumTitle
-          params={{ text: 'How do you want to log in?', classNames: [] }}
-        />
-        <Article params={{ className: styles.card_container }}>
-          <SmallTitle params={{ text: 'User' }} />
-          <div>
-            <Card params={{ userFullName: user?.getFullName() }} />
-          </div>
-        </Article>
-        <Article params={{ className: styles.card_container }}>
-          <SmallTitle params={{ text: 'Companies' }} />
-          <div>
-            {companies.map((company) => (
-              <Card key={company.getUuid()} params={{ company }} />
-            ))}
-          </div>
-        </Article>
-      </Section>
-    </div>
+    <Suspense fallback={null}>
+      <ProfilesClient />
+    </Suspense>
   )
 }
